- renamed ext to tags - changed pom.xml to not use properties when property is referenced once - removed Constatns.java
