+++ opentracing-api/pom.xml
@@ -31,14 +31,13 @@
 
     <properties>
         <main.basedir>${project.basedir}/..</main.basedir>
-        <mockito.version>1.10.19</mockito.version>
     </properties>
 
     <dependencies>
         <dependency>
             <groupId>org.mockito</groupId>
             <artifactId>mockito-all</artifactId>
-            <version>${mockito.version}</version>
+            <version>1.10.19</version>
             <scope>test</scope>
         </dependency>
     </dependencies>
