+++ opentracing-api/src/main/java/io/opentracing/tag/ShortTag.java
@@ -11,7 +11,7 @@
  * or implied. See the License for the specific language governing permissions and limitations under
  * the License.
  */
-package io.opentracing.ext;
+package io.opentracing.tag;
 
 public class ShortTag extends AbstractTag<Short> {
     ShortTag(String key) {
