+++ opentracing-api/src/main/java/io/opentracing/tag/StringTag.java
@@ -11,24 +11,19 @@
  * or implied. See the License for the specific language governing permissions and limitations under
  * the License.
  */
-package io.opentracing.ext;
+package io.opentracing.tag;
 
-import io.opentracing.Span;
-import org.junit.Test;
+public class StringTag extends AbstractTag<String> {
+    StringTag(String key) {
+        super(key);
+    }
 
-import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.verify;
+    @Override
+    public void set(io.opentracing.Span span, String tagValue) {
+        span.setTag(super.key, tagValue);
+    }
 
-public class IntTagTest {
-    @Test
-    public void testSetInt() {
-        Integer value = 7;
-        String key = "expected.key";
-        Span span = mock(Span.class);
-
-        IntTag tag = new IntTag(key);
-        tag.set(span, value);
-
-        verify(span).setTag(key, value);
+    public void set(io.opentracing.Span span, StringTag tag) {
+        span.setTag(super.key, tag.key);
     }
 }
